Recents: Add build.prop config to specify maximum number of recent apps to keep
Allows the user to specify a more "sane" number of recent apps to keep
in the stack instead of the system default value of 100. Value is set
through build.prop key "ro.config.max_recents". Setting this to 0 is
acceptable, and will result in no apps being kept in the recents stack.
Setting this to a value less than 0, greater than 98 or removing the key
will tell the system to use the default value of 50 or 100 depending on
if device is configured as low RAM or not.